+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Animated loading indicator for assistant message generation with spinning Braille pattern animation
+
 ## [0.2.1] - 2025-01-26
 
 ### Fixed